requirements.txt
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
httplib2==0.20.2
idna==3.3
IMDbPY==2021.4.18
lxml==4.7.1
numpy==1.22.0
pandas==1.3.5
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
uritemplate==4.1.1
urllib3==1.26.7
